+++ strom/__init__.py
@@ -1,6 +1,7 @@
 from .api_utils import find_root_dir, get_weather_data, get_price_series
-from .optimization_utils import find_heating_output, compare_output_costs, plot_state
+from .optimization_utils import find_heating_output, compare_output_costs
 from .data_utils import get_temp_price_df, join_data
+from .plot_utils import plot_state
 
 __all__ = [
     'find_root_dir',
